+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Running mypy on SDK resources
   ([#4053](https://github.com/open-telemetry/opentelemetry-python/pull/4053))
+- Drop Final annotation from Enum in semantic conventions
+  ([#4085](https://github.com/open-telemetry/opentelemetry-python/pull/4085))
 
 ## Version 1.26.0/0.47b0 (2024-07-25)